move stuff to util file
